Replacing new hook with old hook.
This new function allows copying old hook functions to a new hook.  This function is helpful because as you don't have class object then you're not able to hook that function into a new hook location. So this will clone the old hook data into a new hook and also will remove the old hook data.